Initial commit message for adding game template files
___________________________________

This repo serves as a template for Java game development using Swing. It includes essential classes for game window management, rendering, animation, and player entity representation.
More features will be added over time to enhance functionality and ease of use. This will be used as the base for future Java game projects.
___________________________________